+++ src/frontend/src/services/bookingCanisterService.ts
@@ -5,8 +5,8 @@ import { createActor, canisterId } from "../../../declarations/booking";
 import { canisterId as reviewCanisterId } from "../../../declarations/review";
 import { canisterId as reputationCanisterId } from "../../../declarations/reputation";
 import { canisterId as serviceCanisterId } from "../../../declarations/service";
-import {canisterId as commissionCanisterId} from "../../../declarations/commission"
-import {canisterId as walletCanisterId} from "../../../declarations/wallet"
+import { canisterId as commissionCanisterId } from "../../../declarations/commission";
+import { canisterId as walletCanisterId } from "../../../declarations/wallet";
 import { Identity } from "@dfinity/agent";
 import type {
   _SERVICE as BookingService,